requirements/requirements.txt
requests
youtube-dl
validators
discord
python-dotenv
asyncio
mypy
mysql-connector-python
pynacl